setup-liboboe - stripped down
- move previous to archived (don't lose download code)
